Fix Helpers::random() argument types mismatch in Match::getRandomMVP()

+++ Osiris/InventoryChanger/ItemGenerator.cpp
@@ -87,7 +87,7 @@ struct Match {
     {
         if (!hasMVPs())
             return ProPlayer{};
-        return mvpPlayers[Helpers::random(0, std::distance(mvpPlayers.begin(), std::ranges::find(mvpPlayers, ProPlayer{})) - 1)];
+        return mvpPlayers[Helpers::random(static_cast<std::size_t>(0), static_cast<std::size_t>(std::distance(mvpPlayers.begin(), std::ranges::find(mvpPlayers, ProPlayer{}))) - 1)];
     }
 };
